Shared BleTester page as component
