+++ SampleClientSide/Pages/BleTester.razor
@@ -1,66 +1,5 @@
 ﻿@page "/"
 
-@inherits BindableBase
-
 <h3>Blazor Web Bluetooth API tester</h3>
 
-<div class="container text-start">
-    <div class="row">
-
-        <div class="col">
-            <div class="form-group">
-                <p>
-                    <button class="btn btn-outline-primary" @onclick="OnGetAvailabilityClicked" disabled="@IsBusy" type="button">Get availability of bluetooth adapter</button>
-                    <button class="btn btn-outline-primary" @onclick="OnGetDeviceClicked" disabled="@IsBusy" type="button">Get latest connected devices</button>
-                </p>
-
-            </div>
-
-            <DeviceRequestComponent Logs="Logs" OnDeviceReceived="OnDeviceReceived" />
-
-            <div class="form-group" style="background: #E8F6F3;">
-
-                @if (Devices.Any())
-                {
-                    @foreach (var device in Devices)
-                    {
-                        <div class="form-group">
-                            <button class="btn btn-outline-danger justify-content-end" @onclick="(() => RemoveDevice(device))" disabled="@IsBusy" type="button">X</button>
-                            <DeviceComponent Device="device" Logs="Logs" />
-                        </div>
-                    }
-                }
-                else
-                {
-                    <p>No devices got yet</p>
-                }
-
-            </div>
-
-        </div>
-
-        <div class="col-3">
-            
-            <p>Logs: </p>
-            @if (Logs != null && Logs.Any())
-            {
-                <div class="border border-warning">
-                    @for (var i = Logs.Count; i > 0; i--)
-                    {
-                        @if (i == Logs.Count)
-                        {
-                            <div class="border border-success">
-                                <p>@Logs[i - 1]</p>
-                            </div>
-
-                        }
-                        else
-                        {
-                            <p>@Logs[i - 1]</p>
-                        }
-                    }
-                </div>
-            }
-        </div>
-    </div>
-</div>
+<BleTesterPageComponent />
